Further source refactoring

+++ src/communication/incoming/room/G_OBJS.h
@@ -14,7 +14,7 @@ void G_OBJS(player *player, incoming_message *message) {
     room *room = player->room_user->room;
     outgoing_message *om = om_create(30); // "@^"
 
-    for (int i = 0; i < list_size(room->public_items); i++) {
+    for (size_t i = 0; i < list_size(room->public_items); i++) {
         item *room_item;
         list_get_at(room->public_items, i, (void*)&room_item);
